Further reduction of the size of the sector TTE tables For default memcheck configuration, 32 bits) this patch decreases by 13.6 MB ie. from 89945856 to 76317696.
Note that the type EClassNo is introduced only for readibility
purpose (and avoid some cast). That does not change the size
of the TTEntry.
The TTEntry size is reduced by using unions and/or Bool on 1 bit.

No performance impact detected (outer callgrind/inner memcheck bz2
on x86 shows a small improvement).



git-svn-id: svn://svn.valgrind.org/valgrind@15054 a5019735-40e9-0310-863c-91ae7b9d1cf9