Patch 7 of 7, improve PPC HW capabiltiy checking.
The patch was submitted by Will Schmidt  (will_schmidt@vnet.ibm.com).

A handful of cosmetic changes to the ppc32 and ppc64 makefiles.
Cleans up some whitespace issues, spaces/tabs/etc,
Rearranges some of the contents so they are logically group, and
more consistent between the 32- and 64- bit versions of the same.

Bugzilla 34979


git-svn-id: svn://svn.valgrind.org/valgrind@15425 a5019735-40e9-0310-863c-91ae7b9d1cf9