Improve mmap MAP_HUGETLB support. This is a follow up to r14682:
When an mmap retry is done without any constraints, the kernel can
place it into free or reservation segments (i.e. anywhere there is no
mapping yet).
In r14682 a sanity check made the hypothesis that the new mapping was
in a free segment, but it does not hold at least on Linux 3.12 and 3.16
on amd64 (tested under Debian).
There is no risk in allowing the mapping to end up in (what was
previously) a reservation at this point, because it is also allowed.

Patch by Guillaume Knispel <xilun0@gmail.com>. Fixes BZ #348269.


git-svn-id: svn://svn.valgrind.org/valgrind@15538 a5019735-40e9-0310-863c-91ae7b9d1cf9