Follow up to fix 376956 syswrap of SNDDRV and DRM_IOCTL_VERSION causing some         addresses to be wrongly marked as addressable
As noted by Ivo, if the syscall fails, then we have a leak.

So, enable the flag SfPostOnFail if we allocate memory.
In the POST ioctl, check that FAILURE only happens for this drm ioctl,
and free the memory for both SUCCESS and FAILURE.
Do the POST_MEM_WRITE only if SUCCESS
        


git-svn-id: svn://svn.valgrind.org/valgrind@16277 a5019735-40e9-0310-863c-91ae7b9d1cf9